+++ App.tsx
@@ -1,12 +1,14 @@
-import "react-native-gesture-handler";
-import React from "react";
-import { StyleSheet } from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
-import { NavigationContainer } from "@react-navigation/native";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import NearbyScreen from "./src/screens/Nearby";
-import ChatScreen from "./src/screens/Chat";
-import ProfileScreen from "./src/screens/Profile";
+import 'react-native-gesture-handler';
+
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { NavigationContainer } from '@react-navigation/native';
+import React from 'react';
+import { StyleSheet } from 'react-native';
+import { SafeAreaView } from 'react-native-safe-area-context';
+
+import { ChatListScreen } from './src/screens/ChatList';
+import NearbyScreen from './src/screens/Nearby';
+import ProfileScreen from './src/screens/Profile';
 
 const Tab = createBottomTabNavigator();
 
@@ -19,7 +21,7 @@ export default function App() {
           initialRouteName="Chat"
         >
           <Tab.Screen name="Nearby" component={NearbyScreen} />
-          <Tab.Screen name="Chat" component={ChatScreen} />
+          <Tab.Screen name="Chat" component={ChatListScreen} />
           <Tab.Screen name="Profile" component={ProfileScreen} />
         </Tab.Navigator>
       </NavigationContainer>
